Attempt at a windows script

+++ build-exiftool.ts
@@ -1,13 +1,23 @@
 import { execSync } from 'node:child_process';
 import { createWriteStream } from 'node:fs';
 import fs from 'node:fs/promises';
-import { tmpdir } from 'node:os';
+import { platform, tmpdir } from 'node:os';
 import path from 'node:path';
 import { Readable } from 'node:stream';
 import { extract } from 'tar';
+import unzipper from 'unzipper';
 
-const exiftoolName = 'Image-ExifTool-13.38';
-const downloadFileName = `${exiftoolName}.tar.gz`;
+const onWindows = platform() === 'win32';
+
+let exiftoolName: string;
+let downloadFileName: string;
+if (onWindows) {
+  exiftoolName = 'exiftool-13.42_64';
+  downloadFileName = `${exiftoolName}.zip`;
+} else {
+  exiftoolName = 'Image-ExifTool-13.42';
+  downloadFileName = `${exiftoolName}.tar.gz`;
+}
 const downloadUrl = `https://sourceforge.net/projects/exiftool/files/${downloadFileName}`;
 
 console.log(`Let's get ${exiftoolName}`);
@@ -19,11 +29,11 @@ if (!res.ok || res.body === null) {
 }
 
 const tempDir = await fs.mkdtemp(tmpdir());
-const tarballPath = path.join(tempDir, downloadFileName);
+const archivePath = path.join(tempDir, downloadFileName);
 
 try {
   const nodeStream = Readable.from(res.body);
-  const fileStream = createWriteStream(tarballPath);
+  const fileStream = createWriteStream(archivePath);
 
   await new Promise<void>((resolve, reject) => {
     nodeStream.pipe(fileStream);
@@ -32,30 +42,46 @@ try {
   });
 
   console.log('Extracting...');
-  await extract({
-    file: tarballPath,
-    cwd: tempDir,
-    gzip: true,
-  });
 
   const extractedPath = path.join(tempDir, exiftoolName);
+  let distPath: string;
 
-  console.log('Building ExifTool...');
-  execSync('perl Makefile.PL', { cwd: extractedPath });
-  execSync('make', { cwd: extractedPath });
+  if (onWindows) {
+    const zipDir = await unzipper.Open.file(archivePath);
+    await zipDir.extract({ path: tempDir });
+    distPath = path.join(extractedPath);
 
-  console.log('Cleaning up build...');
-  const distPath = path.join(extractedPath, 'blib');
-  await fs.rename(
-    path.join(distPath, 'script', 'exiftool'),
-    path.join(distPath, 'exiftool'),
-  );
+    await fs.rename(
+      path.join(distPath, 'exiftool(-k).exe'),
+      path.join(distPath, 'exiftool.exe'),
+    );
+  } else {
+    await extract({
+      file: archivePath,
+      cwd: tempDir,
+      gzip: true,
+    });
 
-  await fs.rm(path.join(distPath, 'arch'), { recursive: true, force: true });
-  await fs.rm(path.join(distPath, 'bin'), { recursive: true, force: true });
-  await fs.rm(path.join(distPath, 'man1'), { recursive: true, force: true });
-  await fs.rm(path.join(distPath, 'man3'), { recursive: true, force: true });
-  await fs.rm(path.join(distPath, 'script'), { recursive: true, force: true });
+    console.log('Building ExifTool...');
+    execSync('perl Makefile.PL', { cwd: extractedPath });
+    execSync('make', { cwd: extractedPath });
+
+    console.log('Cleaning up build...');
+    distPath = path.join(extractedPath, 'blib');
+    await fs.rename(
+      path.join(distPath, 'script', 'exiftool'),
+      path.join(distPath, 'exiftool'),
+    );
+
+    await fs.rm(path.join(distPath, 'arch'), { recursive: true, force: true });
+    await fs.rm(path.join(distPath, 'bin'), { recursive: true, force: true });
+    await fs.rm(path.join(distPath, 'man1'), { recursive: true, force: true });
+    await fs.rm(path.join(distPath, 'man3'), { recursive: true, force: true });
+    await fs.rm(path.join(distPath, 'script'), {
+      recursive: true,
+      force: true,
+    });
+  }
 
   console.log('Putting ExifTool dist in project...');
   const targetDir = path.join('src-tauri', 'resources');
